Create README.md
This project is a clone of Amazon. It is an opportunity for me to learn other aspect of the web development aside my current frontend stack. This was build while learning MERN(MongoDB, Express, React, Node) and React Bootstrap.
This app has all the features of the of the amazon app like, user authentication, add and remove from cart, increase and decrease quantity in card, payment with paystack, e.t.c